feat: 게시글 수정 및 삭제 기능 개선
- 게시글 수정 시 유효성 검사 추가 (작성자 일치 여부 확인)
- 게시글 삭제 시 간단한 예외 처리 로직 추가
- 수정 및 삭제 로직을 서비스 계층으로 이동하여 가독성과 재사용성 향상
- 엔티티 대신 DTO 사용으로 데이터 무결성 강화
- 게시판별 부가 설명 Enum 적용
- 게시글 상세보기 시 조회수 증가 로직 구현
- 게시판 새 글 수 및 전체 글 수 표시 기능 추가
- CSS 수정: 테이블 스타일, 구분선, 사이드바 디자인 개선

+++ src/main/java/com/meatmeet6kae/service/board/BoardService.java
@@ -4,11 +4,15 @@ import com.meatmeet6kae.dto.board.BoardDto;
 import com.meatmeet6kae.entity.board.Board;
 import com.meatmeet6kae.entity.user.User;
 import com.meatmeet6kae.repository.board.BoardRepository;
+import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import org.springframework.ui.Model;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Optional;
 
@@ -30,7 +34,7 @@ public class BoardService {
     }
 
     public List<Board>getBoardsByCategory(String boardCategory){
-        return boardRepository.findByBoardCategory(boardCategory);
+        return boardRepository.findByBoardCategoryOrderByBoardNoDesc(boardCategory);
     }
 
     // 게시글 생성
@@ -46,20 +50,40 @@ public class BoardService {
     }
 
     public Board getBoardByBoardNo(int boardNo) {
-        return boardRepository.findByBoardNo(boardNo).orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없습니다."));
+        return boardRepository.findByBoardNo(boardNo).orElse(null);
     }
 
-    public Board updateBoard(Board board) {
+    // 게시글 수정
+    public Board updateBoard(BoardDto boardDto) {
 
-        Board boards = getBoardByBoardNo(board.getBoardNo());
-        boards.setTitle(board.getTitle());
-        boards.setContent(board.getContent());
+        Board boards = getBoardByBoardNo(boardDto.getBoardNo());
+
+        // 게시글 수정 내용
+        boards.setTitle(boardDto.getTitle());
+        boards.setContent(boardDto.getContent());
         boards.setUpdateDate(LocalDateTime.now());
         return boardRepository.save(boards);
     }
 
-    public void deleteBoard(int boardNo){
+    //게시글 삭제
+    public void deleteBoard(int boardNo, User user){
         Board board = getBoardByBoardNo(boardNo);
+
+        // 유효성 검사: 게시글이 없거나 작성자와 로그인한 사용자 비교
+        if (board == null) {
+            return;
+        }
+        if (!board.getUser().getLoginId().equals(user.getLoginId())){
+            return;
+        }
         boardRepository.delete(board);
+    }
+
+    // 게시글 조회수 계산
+    @Transactional
+    public void updateViewCount(int boardNo){
+        Board board = getBoardByBoardNo(boardNo);
+
+        board.setViewCount(board.getViewCount()+1);
     }
 }
